server/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
google-auth>=2.25
PyJWT>=2.8
requests>=2.31
email-validator>=2.1
